+++ functions.php
@@ -9,7 +9,7 @@
 defined( 'ABSPATH' ) || exit;
 
 if ( ! defined( 'MOGHADAM_VERSION' ) ) {
-	define( 'MOGHADAM_VERSION', '1.3.0' );
+	define( 'MOGHADAM_VERSION', '1.4.0' );
 }
 
 if ( ! defined( 'MOGHADAM_DIR' ) ) {
@@ -25,7 +25,6 @@ require_once MOGHADAM_DIR . '/inc/setup.php';
 require_once MOGHADAM_DIR . '/inc/enqueue.php';
 require_once MOGHADAM_DIR . '/inc/layout.php';
 require_once MOGHADAM_DIR . '/inc/canvas.php';
-require_once MOGHADAM_DIR . '/inc/style-guide.php';
 require_once MOGHADAM_DIR . '/inc/template-tags.php';
 require_once MOGHADAM_DIR . '/inc/icons.php';
 require_once MOGHADAM_DIR . '/inc/home/home-settings.php';
@@ -35,8 +34,11 @@ require_once MOGHADAM_DIR . '/inc/design-tags.php';
 require_once MOGHADAM_DIR . '/inc/social.php';
 require_once MOGHADAM_DIR . '/inc/case-studies.php';
 require_once MOGHADAM_DIR . '/inc/portfolio.php';
+require_once MOGHADAM_DIR . '/inc/post-sidebar.php';
 require_once MOGHADAM_DIR . '/inc/customizer.php';
 require_once MOGHADAM_DIR . '/inc/home/customizer-home.php';
+
+require_once MOGHADAM_DIR . '/inc/post-sidebar-admin.php';
 
 if ( is_admin() ) {
 	require_once MOGHADAM_DIR . '/inc/settings.php';